+++ server/cmd/cli/gram/api/deployments.go
@@ -3,8 +3,6 @@ package api
 import (
 	"context"
 	"fmt"
-	"log"
-	"strings"
 
 	"github.com/speakeasy-api/gram/server/cmd/cli/gram/env"
 
@@ -21,26 +19,6 @@ func NewDeploymentsClient() *DeploymentsClient {
 	return &DeploymentsClient{
 		client: newDeploymentClient(),
 	}
-}
-
-func (c *DeploymentsClient) ListDeployments(
-	apiKey string,
-	projectSlug string,
-) *deployments.ListDeploymentResult {
-	ctx := context.Background()
-	payload := &deployments.ListDeploymentsPayload{
-		ApikeyToken:      &apiKey,
-		ProjectSlugInput: &projectSlug,
-		SessionToken:     nil,
-		Cursor:           nil,
-	}
-
-	result, err := c.client.ListDeployments(ctx, payload)
-	if err != nil {
-		log.Fatalf("Error calling ListDeployments: %v", err)
-	}
-
-	return result
 }
 
 // DeploymentCreator represents a request for creating a deployment
@@ -81,24 +59,10 @@ func (c *DeploymentsClient) CreateDeployment(
 
 	result, err := c.client.CreateDeployment(ctx, payload)
 	if err != nil {
-		enhancedErr := enhanceDeploymentError(err)
-		return nil, fmt.Errorf("failed to create deployment: %w", enhancedErr)
+		return nil, fmt.Errorf("failed to create deployment: %w", err)
 	}
 
 	return result, nil
-}
-
-// enhanceDeploymentError provides more context for deployment errors,
-// especially decode errors that may indicate server issues.
-func enhanceDeploymentError(err error) error {
-	errStr := err.Error()
-
-	// Check if this is a decode error that suggests the server returned HTML instead of JSON
-	if strings.Contains(errStr, "can't decode") && strings.Contains(errStr, "text/html") {
-		return fmt.Errorf("%w\n\nThis error typically occurs when the server returns an HTML error page (e.g., 500 error) instead of the expected JSON response. Check server logs or try again later", err)
-	}
-
-	return err
 }
 
 func newDeploymentClient() *deployments.Client {
