+++ server/cmd/cli/gram/log/log.go
@@ -1,0 +1,8 @@
+package log
+
+import (
+	"log/slog"
+	"os"
+)
+
+var L = slog.New(slog.NewTextHandler(os.Stderr, nil))
