+++ server/cmd/cli/gram/main.go
@@ -1,14 +1,15 @@
 package main
 
 import (
-	"log"
 	"os"
 
 	"github.com/speakeasy-api/gram/server/cmd/cli/gram/app"
+	"github.com/speakeasy-api/gram/server/cmd/cli/gram/log"
 )
 
 func main() {
 	if err := app.NewCLI().Run(os.Args); err != nil {
-		log.Fatal(err)
+		log.L.Error("CLI failed", "error", err)
+		os.Exit(1)
 	}
 }
